+++ src/app/core/components/footer/footer.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from "@angular/core";
-import { User } from "src/app/features/public/models/User";
+import { User } from "src/app/features/models/User";
 import { UserStatusService } from "../../services/user-status.service";
 
 @Component({
