+++ src/app/features/backoffice/pages/activity-form/activity-form.component.ts
@@ -4,8 +4,8 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { CkeditorService } from 'src/app/core/services/ckeditor.service';
 import { HttpService } from 'src/app/core/services/http.service';
 import { UserStatusService } from 'src/app/core/services/user-status.service';
-import { Activity } from 'src/app/features/public/models/Activity';
-import { ImageFile } from 'src/app/features/public/models/ImageFile';
+import { Activity } from 'src/app/features/models/Activity';
+import { ImageFile } from 'src/app/features/models/ImageFile';
 import { UrlImageValidator } from 'src/app/shared/utils/url-image.validator';
 import { environment } from 'src/environments/environment';
 import Swal from'sweetalert2';
